+++ scripts/danger-dts.ts
@@ -9,8 +9,7 @@ const createDTS = () => {
 //
 
 import { Octokit as GitHub } from "@octokit/rest"
-import { Gitlab, Types } from "@gitbeaker/node"
-import { Types as CoreTypes } from "@gitbeaker/core/dist"
+import type * as Types from "@gitbeaker/rest"
 import { File } from "parse-diff"
 
 `